bugshield_Web/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
bcrypt==5.0.0
PyJWT==2.10.1
email-validator==2.2.0
requests==2.32.5
reportlab==4.2.5
pydantic==2.10.3
chardet<6
